fix(skene): align session history contract (#4911)
Gate-Passed: kanon 0.1.0

+++ crates/theatron/proskenion/src/components/session_tabs.rs
@@ -81,7 +81,7 @@ type TabSnapshot = (
     NousId,
     Option<String>,
     Option<SessionId>,
-    Option<u32>,
+    Option<i64>,
     String,
     bool,
     bool,
